Replaced join on ship_flags with locate "nebelscan" as ship_flags is empty as of now.
Exclude landed ships from displaying in starmap (F's for friendly, neutral or enemy fleets)

Bugfix in cargo: when the item repository was introduced, the logic to check whether the item is of the desired type got broken, due to it relying on inheritance.
(Also added a new method and called that one in WerftGUI instead)

Small fix in Starmap -> goto windows is now closed when using goto location.